Complete Aurora PostgreSQL operational toolkit
Finish all workflow categories, add generated category indexes, and harden read-only diagnostics using live Aurora PostgreSQL 17.7 validation feedback. Guard unsupported WAL instrumentation, permission-restricted statistics, optional extensions, and parameterized workflows while retaining community PostgreSQL paths.

+++ database-health/post-deployment-check/scripts/01_error_and_rollback_rates.sql
@@ -57,7 +57,7 @@ SELECT
     now() - stats_reset                                           AS counting_window,
     round(
         (xact_commit + xact_rollback)::numeric /
-        NULLIF(EXTRACT(epoch FROM (now() - stats_reset)), 0),
+        NULLIF(EXTRACT(epoch FROM (now() - stats_reset))::numeric, 0),
         2
     )                                                             AS avg_txn_per_second
 FROM pg_stat_database
